rename: utils 파일 타입 변경 .js->.ts
