+++ json/savedGame.json
@@ -21,7 +21,6 @@
             },
             { "squares": ["X", "O", null, null, null, null, null, null, null] },
             { "squares": ["X", "O", null, null, null, null, null, null, "X"] }
-        ],
-        "date": "22년 10월 28일 오전 2:25:42"
+        ]
     }
 }
