+++ types/game.ts
@@ -2,7 +2,6 @@ export interface GameData {
     stepNumber: number;
     xIsNext: boolean;
     history: SquaresType[];
-    date: string | null;
 }
 
 export interface SquaresType {
