+++ utils/calculateWinner.ts
@@ -1,6 +1,6 @@
-import useHistory from './hooks/useHistory';
+import { SquaresType } from '../types/game';
 
-export default function calculateWinner(squares) {
+export default function calculateWinner(squares: (string | null)[]) {
     const lines = [
         [0, 1, 2],
         [3, 4, 5],
